# openapi.json
{"openapi":"3.1.0","info":{"title":"AI Work System Scaffold","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}},"/intake/brief":{"post":{"summary":"Intake Brief","operationId":"intake_brief_intake_brief_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntakeBriefRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntakeResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orchestrator/run":{"post":{"summary":"Orchestrator Run","operationId":"orchestrator_run_orchestrator_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrchestratorRunRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrchestrationResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orchestrator/resume/approval":{"post":{"summary":"Resume Approval","operationId":"resume_approval_orchestrator_resume_approval_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApprovalResumeRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrchestrationResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orchestrator/approval/reject":{"post":{"summary":"Reject Approval","operationId":"reject_approval_orchestrator_approval_reject_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApprovalRejectRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrchestrationResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orchestrator/resume/revision":{"post":{"summary":"Resume Revision","operationId":"resume_revision_orchestrator_resume_revision_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RevisionResumeRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrchestrationResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orchestrator/replanning/start":{"post":{"summary":"Start Replanning","operationId":"start_replanning_orchestrator_replanning_start_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplanningStartRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrchestrationResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/audit":{"get":{"summary":"Project Audit","operationId":"project_audit_projects__project_id__audit_get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectAudit"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ActorContext":{"properties":{"actor_id":{"type":"string","title":"Actor Id"},"actor_role":{"$ref":"#/components/schemas/ActorRole"},"actor_type":{"$ref":"#/components/schemas/ActorType","default":"human"}},"type":"object","required":["actor_id","actor_role"],"title":"ActorContext"},"ActorRole":{"type":"string","enum":["owner","operator","approver","admin","viewer"],"title":"ActorRole"},"ActorType":{"type":"string","enum":["human","service","system"],"title":"ActorType"},"ApprovalActionType":{"type":"string","enum":["external_api_send","destructive_change","bulk_modify","production_affecting_change"],"title":"ApprovalActionType"},"ApprovalRejectRequest":{"properties":{"project_id":{"type":"string","minLength":1,"title":"Project Id"},"rejected_actions":{"items":{"type":"string"},"type":"array","title":"Rejected Actions"},"actor":{"anyOf":[{"$ref":"#/components/schemas/ActorContext"},{"type":"null"}]},"reason":{"type":"string","minLength":1,"title":"Reason"},"note":{"type":"string","title":"Note","default":""}},"type":"object","required":["project_id","reason"],"title":"ApprovalRejectRequest"},"ApprovalRequest":{"properties":{"id":{"type":"string","title":"Id"},"action_type":{"$ref":"#/components/schemas/ApprovalActionType"},"status":{"$ref":"#/components/schemas/ApprovalStatus","default":"pending"},"reason":{"type":"string","title":"Reason"},"requested_by":{"type":"string","title":"Requested By","default":"system"},"decision_note":{"type":"string","title":"Decision Note","default":""}},"type":"object","required":["id","action_type","reason"],"title":"ApprovalRequest"},"ApprovalResumeRequest":{"properties":{"project_id":{"type":"string","minLength":1,"title":"Project Id"},"approved_actions":{"items":{"type":"string"},"type":"array","title":"Approved Actions"},"actor":{"anyOf":[{"$ref":"#/components/schemas/ActorContext"},{"type":"null"}]},"note":{"type":"string","title":"Note","default":""},"trend_provider":{"type":"string","title":"Trend Provider","default":"mock"}},"type":"object","required":["project_id"],"title":"ApprovalResumeRequest"},"ApprovalStatus":{"type":"string","enum":["pending","approved","rejected"],"title":"ApprovalStatus"},"Artifact":{"properties":{"id":{"type":"string","title":"Id"},"task_id":{"type":"string","title":"Task Id"},"artifact_type":{"type":"string","title":"Artifact Type"},"content":{"additionalProperties":true,"type":"object","title":"Content"}},"type":"object","required":["id","task_id","artifact_type"],"title":"Artifact"},"Checkpoint":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"approved":{"type":"boolean","title":"Approved","default":false},"approver":{"type":"string","title":"Approver","default":"human"},"note":{"type":"string","title":"Note","default":""}},"type":"object","required":["id","name"],"title":"Checkpoint"},"Department":{"type":"string","enum":["research","design","build","review","trend"],"title":"Department"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HistoryEvent":{"properties":{"event_type":{"$ref":"#/components/schemas/HistoryEventType"},"actor":{"type":"string","title":"Actor","default":"system"},"actor_role":{"$ref":"#/components/schemas/ActorRole","default":"admin"},"actor_type":{"$ref":"#/components/schemas/ActorType","default":"system"},"timestamp":{"type":"string","title":"Timestamp"},"reason":{"type":"string","title":"Reason","default":""},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata"}},"type":"object","required":["event_type"],"title":"HistoryEvent"},"HistoryEventType":{"type":"string","enum":["authentication_succeeded","authentication_failed","actor_resolved","state_transition","approval_requested","approval_approved","approval_rejected","authorization_granted","authorization_failed","policy_override_applied","revision_requested","replanning_started","resume_triggered","task_status_changed"],"title":"HistoryEventType"},"IntakeBriefRequest":{"properties":{"user_request":{"type":"string","minLength":1,"title":"User Request"}},"type":"object","required":["user_request"],"title":"IntakeBriefRequest"},"IntakeResult":{"properties":{"brief":{"$ref":"#/components/schemas/ProjectBrief"},"missing_fields":{"items":{"type":"string"},"type":"array","title":"Missing Fields"},"clarifying_questions":{"items":{"type":"string"},"type":"array","title":"Clarifying Questions"}},"type":"object","required":["brief"],"title":"IntakeResult"},"OrchestrationResult":{"properties":{"record":{"$ref":"#/components/schemas/ProjectRecord"},"summary":{"$ref":"#/components/schemas/ProjectSummary"}},"type":"object","required":["record","summary"],"title":"OrchestrationResult"},"OrchestratorRunRequest":{"properties":{"brief":{"$ref":"#/components/schemas/ProjectBrief"},"project_policy":{"anyOf":[{"$ref":"#/components/schemas/ProjectPolicy-Input"},{"type":"null"}]},"trend_provider":{"type":"string","title":"Trend Provider","default":"mock"},"approved_actions":{"items":{"type":"string"},"type":"array","title":"Approved Actions"},"simulate_review_failure":{"type":"boolean","title":"Simulate Review Failure","default":false}},"type":"object","required":["brief"],"title":"OrchestratorRunRequest"},"Project":{"properties":{"id":{"type":"string","title":"Id"},"brief":{"$ref":"#/components/schemas/ProjectBrief"},"policy":{"$ref":"#/components/schemas/ProjectPolicy-Output"},"status":{"$ref":"#/components/schemas/ProjectStatus","default":"draft"}},"type":"object","required":["id","brief"],"title":"Project"},"ProjectAudit":{"properties":{"project_id":{"type":"string","title":"Project Id"},"status":{"$ref":"#/components/schemas/ProjectStatus"},"history":{"items":{"type":"string"},"type":"array","title":"History"},"events":{"items":{"$ref":"#/components/schemas/HistoryEvent"},"type":"array","title":"Events"},"approvals":{"items":{"$ref":"#/components/schemas/ApprovalRequest"},"type":"array","title":"Approvals"},"reviews":{"items":{"$ref":"#/components/schemas/Review"},"type":"array","title":"Reviews"},"checkpoints":{"items":{"$ref":"#/components/schemas/Checkpoint"},"type":"array","title":"Checkpoints"}},"type":"object","required":["project_id","status"],"title":"ProjectAudit"},"ProjectBrief":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"objective":{"type":"string","title":"Objective"},"scope":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scope"},"constraints":{"items":{"type":"string"},"type":"array","title":"Constraints"},"success_criteria":{"items":{"type":"string"},"type":"array","title":"Success Criteria"},"deadline":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Deadline"},"stakeholders":{"items":{"type":"string"},"type":"array","title":"Stakeholders"},"assumptions":{"items":{"type":"string"},"type":"array","title":"Assumptions"},"raw_request":{"type":"string","title":"Raw Request"}},"type":"object","required":["objective","raw_request"],"title":"ProjectBrief"},"ProjectPolicy-Input":{"properties":{"project_owner_actor_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Project Owner Actor Id"},"strict_mode":{"type":"boolean","title":"Strict Mode","default":false},"action_rules":{"additionalProperties":{"$ref":"#/components/schemas/ProjectPolicyActionRule"},"type":"object","title":"Action Rules"}},"type":"object","title":"ProjectPolicy"},"ProjectPolicy-Output":{"properties":{"project_owner_actor_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Project Owner Actor Id"},"strict_mode":{"type":"boolean","title":"Strict Mode","default":false},"action_rules":{"additionalProperties":{"$ref":"#/components/schemas/ProjectPolicyActionRule"},"type":"object","title":"Action Rules"}},"type":"object","title":"ProjectPolicy"},"ProjectPolicyActionRule":{"properties":{"allowed_roles":{"items":{"$ref":"#/components/schemas/ActorRole"},"type":"array","title":"Allowed Roles"},"allowed_actor_ids":{"items":{"type":"string"},"type":"array","title":"Allowed Actor Ids"}},"type":"object","title":"ProjectPolicyActionRule"},"ProjectRecord":{"properties":{"project":{"$ref":"#/components/schemas/Project"},"tasks":{"items":{"$ref":"#/components/schemas/Task"},"type":"array","title":"Tasks"},"artifacts":{"items":{"$ref":"#/components/schemas/Artifact"},"type":"array","title":"Artifacts"},"reviews":{"items":{"$ref":"#/components/schemas/Review"},"type":"array","title":"Reviews"},"checkpoints":{"items":{"$ref":"#/components/schemas/Checkpoint"},"type":"array","title":"Checkpoints"},"approvals":{"items":{"$ref":"#/components/schemas/ApprovalRequest"},"type":"array","title":"Approvals"},"history":{"items":{"type":"string"},"type":"array","title":"History"},"events":{"items":{"$ref":"#/components/schemas/HistoryEvent"},"type":"array","title":"Events"}},"type":"object","required":["project"],"title":"ProjectRecord"},"ProjectStatus":{"type":"string","enum":["draft","intake_pending","ready_for_planning","in_progress","waiting_approval","review_failed","revision_requested","completed"],"title":"ProjectStatus"},"ProjectSummary":{"properties":{"project_id":{"type":"string","title":"Project Id"},"status":{"$ref":"#/components/schemas/ProjectStatus"},"completed_tasks":{"type":"integer","title":"Completed Tasks"},"artifact_count":{"type":"integer","title":"Artifact Count"},"next_steps":{"items":{"type":"string"},"type":"array","title":"Next Steps"}},"type":"object","required":["project_id","status","completed_tasks","artifact_count"],"title":"ProjectSummary"},"ReplanningStartRequest":{"properties":{"project_id":{"type":"string","minLength":1,"title":"Project Id"},"actor":{"anyOf":[{"$ref":"#/components/schemas/ActorContext"},{"type":"null"}]},"note":{"type":"string","title":"Note","default":""},"trend_provider":{"type":"string","title":"Trend Provider","default":"mock"},"approved_actions":{"items":{"type":"string"},"type":"array","title":"Approved Actions"},"reset_downstream_tasks":{"type":"boolean","title":"Reset Downstream Tasks","default":true}},"type":"object","required":["project_id"],"title":"ReplanningStartRequest"},"Review":{"properties":{"id":{"type":"string","title":"Id"},"task_id":{"type":"string","title":"Task Id"},"verdict":{"type":"string","title":"Verdict"},"findings":{"items":{"type":"string"},"type":"array","title":"Findings"}},"type":"object","required":["id","task_id","verdict"],"title":"Review"},"RevisionResumeMode":{"type":"string","enum":["replanning","rebuilding","rereview"],"title":"RevisionResumeMode"},"RevisionResumeRequest":{"properties":{"project_id":{"type":"string","minLength":1,"title":"Project Id"},"resume_mode":{"$ref":"#/components/schemas/RevisionResumeMode"},"actor":{"anyOf":[{"$ref":"#/components/schemas/ActorContext"},{"type":"null"}]},"reason":{"type":"string","title":"Reason","default":""},"trend_provider":{"type":"string","title":"Trend Provider","default":"mock"},"approved_actions":{"items":{"type":"string"},"type":"array","title":"Approved Actions"}},"type":"object","required":["project_id","resume_mode"],"title":"RevisionResumeRequest"},"Task":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"department":{"$ref":"#/components/schemas/Department"},"status":{"$ref":"#/components/schemas/TaskStatus","default":"pending"},"depends_on":{"items":{"type":"string"},"type":"array","title":"Depends On"},"note":{"type":"string","title":"Note","default":""}},"type":"object","required":["id","title","department"],"title":"Task"},"TaskStatus":{"type":"string","enum":["pending","in_progress","waiting_approval","revision_requested","done","blocked"],"title":"TaskStatus"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}